split date parsing and intepretation
