+++ astronomy/astronomy.vcxproj
@@ -203,8 +203,8 @@
     <Text Include="initial_state_two_bodies_elliptical_test.proto.txt" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="date.hpp" />
-    <ClInclude Include="date_body.hpp" />
+    <ClInclude Include="calendar.hpp" />
+    <ClInclude Include="calendar_body.hpp" />
     <ClInclude Include="time_scales.hpp" />
     <ClInclude Include="time_scales_body.hpp" />
     <ClInclude Include="eop_c04.generated.h" />
@@ -219,6 +219,7 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="calendar_test.cpp" />
     <ClCompile Include="time_scales_test.cpp" />
     <ClCompile Include="lunar_eclipse_test.cpp" />
     <ClCompile Include="mercury_perihelion_test.cpp" />
